+++ ru.taximaxim.codekeeper.core/src/main/java/ru/taximaxim/codekeeper/core/schema/AbstractDatabase.java
@@ -241,6 +241,7 @@ public abstract class AbstractDatabase extends PgStatement implements IDatabase 
         });
 
         overrides.addAll(lib.getOverrides());
+        lib.getObjReferences().entrySet().forEach(e -> objReferences.putIfAbsent(e.getKey(), e.getValue()));
     }
 
     protected void addOverride(PgOverride override) {